+++ Dungeon/Assets/Scripts/11.HUD/StatBar/UI_StatBarImage.cs
@@ -4,22 +4,22 @@ using UnityEngine.UI;
 public class UI_StatBarImage : UI_StatBar
 {
     private const float MaxValueCoef = 0.87f;
-    private int _maxValue = 100;
+    private float _maxValue = 100;
     
     [SerializeField] private Image icon;
     [SerializeField] private Image slider;
     [SerializeField] private TextMeshProUGUI valueText;
 
-    public override void SetStat(int newValue)
+    public override void SetStat(float newValue)
     {
-        float value = (float)newValue / _maxValue;
+        float value = newValue / _maxValue;
         slider.fillAmount = value * MaxValueCoef;
         valueText.text = (int)(value * 100f) + "%";
         slider.color = GetColorGradient(_maxValue);
         icon.color = GetColorGradient(_maxValue);
     }
 
-    public override void SetMaxStat(int maxValue)
+    public override void SetMaxStat(float maxValue)
     {
         _maxValue = maxValue;
         slider.fillAmount = MaxValueCoef;
